Fixed a bug with the removal of selected codes
event.target attribute was being used in the handleRemoveSelectedCode() method to determine the ID of the removed
ICD code. Changed the method to use event.currentTarget attribute instead. event.currentTarget always refers to
the element the event handler has been attached to as opposed to event.target which identifies the element on
which the event occurred.